requirements.txt
fastapi
uvicorn
SQLAlchemy
asyncpg
alembic
pydantic
python-dotenv
psycopg2-binary
annotated-types
anyio
async-timeout
bcrypt
certifi
cffi
click
cryptography
dnspython
ecdsa
email_validator
exceptiongroup
fastapi-cli
greenlet
h11
httpcore
httptools
httpx
idna
importlib_metadata
importlib_resources
Jinja2
Mako
markdown-it-py
MarkupSafe
mdurl
orjson
passlib
pyasn1
pycparser
pydantic_core
Pygments
python-jose
python-multipart
PyYAML
rich
rsa
shellingham
six
sniffio
starlette
typer
typing_extensions
ujson
uvloop
watchfiles
websockets
zipp
aioredis
fastapi-limiter